// src/app/openapi/generated-angular-sdk/model/user.ts
/**
 * Fridge API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface User { 
    id?: number;
    email: string;
    username: string;
    password_hash: string;
    family_id: number;
}